separating search logic
